Add scroll area and move GML file selection
