+++ gml_application_schema_toolbox/gui/import_xml_panel.py
@@ -2,6 +2,8 @@
 
 import os
 from qgis.PyQt import uic
+from qgis.PyQt.QtCore import pyqtSlot
+
 
 WIDGET, BASE = uic.loadUiType(os.path.join(
     os.path.dirname(__file__), '..', 'ui', 'import_xml_panel.ui'))
@@ -12,3 +14,12 @@ class ImportXmlPanel(BASE, WIDGET):
     def __init__(self, parent=None):
         super(ImportXmlPanel, self).__init__(parent)
         self.setupUi(self)
+
+    @pyqtSlot()
+    def on_gmlPathButton_clicked(self):
+        path, filter = QFileDialog.getOpenFileName(self,
+            self.tr("Open GML file"),
+            data_folder,
+            self.tr("GML files (*.gml *.xml)"))
+        if path:
+            self.gmlPathLineEdit.setText(path)
